/comment endpoint throws internal server error
-  need to work on a fix

+++ src/main/java/com/example/twittertrial/Service/CommentService.java
@@ -1,39 +1,74 @@
 package com.example.twittertrial.Service;
 
+import com.example.twittertrial.DTO.CommentDto;
 import com.example.twittertrial.Entity.Comment;
+import com.example.twittertrial.Entity.Post;
+import com.example.twittertrial.Entity.User;
 import com.example.twittertrial.Repository.CommentRepository;
+import com.example.twittertrial.Repository.PostRepository;
+import com.example.twittertrial.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public CommentService(CommentRepository commentRepository) {
+    public CommentService(CommentRepository commentRepository, PostRepository postRepository, UserRepository userRepository) {
         this.commentRepository = commentRepository;
+        this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
-    public Comment saveComment(Comment comment) {
-        return commentRepository.save(comment);
+    public String createComment(CommentDto commentDto) {
+        Optional<Post> optionalPost = postRepository.findById(commentDto.getPostId());
+        if (optionalPost.isPresent()) {
+            Optional<User> optionalUser = userRepository.findById(commentDto.getUserId());
+            if (optionalUser.isPresent()) {
+                Comment comment = new Comment();
+                comment.setCommentBody(commentDto.getCommentBody());
+                comment.setCommentCreator(commentDto.getCommentCreator());
+                comment.setPost(optionalPost.get());
+                comment.setUser(optionalUser.get());
+                commentRepository.save(comment);
+                return "Comment created successfully";
+            } else {
+                return "User does not exist";
+            }
+        } else {
+            return "Post does not exist";
+        }
     }
 
-    public Comment getCommentById(Long commentId) {
-        return commentRepository.findById(commentId)
-                .orElseThrow(() -> new CommentNotFoundException("Comment not found with id: " + commentId));
+    public Optional<Comment> getCommentById(Long commentId) {
+        return commentRepository.findById(commentId);
     }
 
-    public Comment updateComment(Long commentId, Comment updatedComment) {
-        Comment existingComment = getCommentById(commentId);
-        // Update existingComment properties with updatedComment properties
-        existingComment.setId(Long.valueOf(updatedComment.getCommentBody()));
-        // Update other properties as needed
-
-        return commentRepository.save(existingComment);
+    public String editComment(Long commentId, CommentDto commentDto) {
+        Optional<Comment> optionalComment = commentRepository.findById(commentId);
+        if (optionalComment.isPresent()) {
+            Comment comment = optionalComment.get();
+            comment.setCommentBody(commentDto.getCommentBody());
+            commentRepository.save(comment);
+            return "Comment edited successfully";
+        } else {
+            return "Comment does not exist";
+        }
     }
 
-    public void deleteComment(Long commentId) {
-        commentRepository.deleteById(commentId);
+    public String deleteComment(Long commentId) {
+        Optional<Comment> optionalComment = commentRepository.findById(commentId);
+        if (optionalComment.isPresent()) {
+            commentRepository.deleteById(commentId);
+            return "Comment deleted";
+        } else {
+            return "Comment does not exist";
+        }
     }
 }
